refactor(blorktools): js to scripts
