+++ packages/blorktools/src/asset_debugger/main.js
@@ -14,7 +14,7 @@ import { initState } from './core/state.js';
 import { initUiManager } from './ui/ui-manager.js';
 import { setupDropzones } from './ui/file-handler.js';
 import { initAtlasPanel } from './ui/scripts/atlas-panel.js';
-import { initUvPanel } from './ui/uv-panel.js';
+import { initUvPanel } from './ui/scripts/uv-panel.js';
 
 /**
  * Initialize the asset debugger application
