--- packages/blorktools/src/asset_debugger/ui/axis-indicator.html
@@ -1,20 +1,0 @@
-<!-- Axis Indicator Settings Tab -->
-<div id="axis-settings" class="settings-tab-content">
-    <div class="settings-group">
-        <div class="settings-option">
-            <label for="axis-indicator-type">Display Mode</label>
-            <select id="axis-indicator-type" class="settings-select">
-                <option value="disabled" selected>Disabled</option>
-                <option value="windowed">Windowed</option>
-                <option value="embedded">Embedded</option>
-            </select>
-        </div>
-        <div id="embedded-axis-settings" class="settings-option" style="display: none;">
-            <label for="axis-intensity">Background Intensity</label>
-            <div class="slider-container">
-                <input type="range" id="axis-intensity" min="0.1" max="1.0" step="0.1" value="0.7" class="settings-slider">
-                <span id="axis-intensity-value">0.7</span>
-            </div>
-        </div>
-    </div>
-</div>
