--- packages/blorktools/src/asset_debugger/ui/mesh-panel.html
@@ -1,5 +1,0 @@
-<!-- Mesh Panel Component -->
-<div id="mesh-tab" class="tab-content active">
-    <h3>Mesh Visibility</h3>
-    <div id="mesh-groups"></div>
-</div>
