+++ packages/blorktools/src/asset_debugger/ui/pages/settings-modal.html
@@ -22,7 +22,7 @@
                         <label for="axis-indicator-type">Type:</label>
                         <select id="axis-indicator-type" class="settings-select">
                             <option value="disabled">Disabled</option>
-                            <option value="corner">Corner</option>
+                            <option value="corner">Windowed</option>
                             <option value="embedded">Embedded</option>
                         </select>
                     </div>
