+++ packages/blorktools/src/asset_debugger/ui/scripts/uv-panel.js
@@ -3,8 +3,8 @@
  * 
  * This module handles UV mapping visualization and controls.
  */
-import { getState, updateState } from '../core/state.js';
-import { updateUvRegion } from './scripts/atlas-panel.js';
+import { getState, updateState } from '../../core/state.js';
+import { updateUvRegion } from './atlas-panel.js';
 
 // DOM elements
 let uvInfoContainer = null;
