--- packages/blorktools/src/asset_debugger/ui/uv-panel.html
@@ -1,50 +1,0 @@
-<!-- UV Panel Component -->
-<div id="uv-tab" class="tab-content">
-    <h3>UV Channel Controls</h3>
-    <div id="uv-content">
-        <div id="uv-info-container" class="uv-info-container">
-            <p>No model loaded or no UV data available.</p>
-            <p>Load a model to view UV information.</p>
-        </div>
-        <div id="uv-manual-controls" style="display: none;">
-            <h4>Manual UV Mapping Controls</h4>
-            <div class="uv-control-group">
-                <label>UV Offset (0-1):</label>
-                <div class="uv-control-row">
-                    <span>X:</span>
-                    <input type="number" id="uv-offset-x" min="0" max="1" step="0.01" value="0">
-                </div>
-                <div class="uv-control-row">
-                    <span>Y:</span>
-                    <input type="number" id="uv-offset-y" min="0" max="1" step="0.01" value="0">
-                </div>
-            </div>
-            <div class="uv-control-group">
-                <label>UV Scale (0-1):</label>
-                <div class="uv-control-row">
-                    <span>W:</span>
-                    <input type="number" id="uv-scale-w" min="0.01" max="1" step="0.01" value="1">
-                </div>
-                <div class="uv-control-row">
-                    <span>H:</span>
-                    <input type="number" id="uv-scale-h" min="0.01" max="1" step="0.01" value="1">
-                </div>
-            </div>
-            <div class="uv-control-group">
-                <label>Predefined Segments:</label>
-                <select id="uv-predefined-segments">
-                    <option value="0">Full texture (1×1)</option>
-                    <option value="1">Top-left quarter (1/2×1/2)</option>
-                    <option value="2">Top-right quarter (1/2×1/2)</option>
-                    <option value="3">Bottom-left quarter (1/2×1/2)</option>
-                    <option value="4">Bottom-right quarter (1/2×1/2)</option>
-                    <option value="5">Top-left ninth (1/3×1/3)</option>
-                    <option value="6">Top-center ninth (1/3×1/3)</option>
-                    <option value="7">Top-right ninth (1/3×1/3)</option>
-                    <option value="8">Middle-left ninth (1/3×1/3)</option>
-                </select>
-            </div>
-            <div class="uv-note">Changes auto-apply immediately</div>
-        </div>
-    </div>
-</div>
